LayoutTests/webaudio/Analyser/handle-silent-inputs-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "connected" 
PASS Executing "auto-pull" 
PASS Executing "timing" 
PASS Audit report 
PASS > [connected] Test handling of silent inputs 
FAIL X 0: Analyser time data at time 0.5: Expected 0 for all values but found 64 unexpected values: 
	Index	Actual
	[0]	0.9980266094207764
	[1]	0.9939585328102112
	[2]	0.960288405418396
	[3]	0.8980192542076111
	...and 60 more errors. assert_true: expected true got false
FAIL < [connected] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [auto-pull] Test handling of silent inputs 
FAIL X 1: Analyser time data at time 0.5: Expected 0 for all values but found 64 unexpected values: 
	Index	Actual
	[0]	0.9980266094207764
	[1]	0.9939585328102112
	[2]	0.960288405418396
	[3]	0.8980192542076111
	...and 60 more errors. assert_true: expected true got false
FAIL < [auto-pull] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [timing] Test shifting in of zeroes after source has stopped 
PASS   At time 0.125 Analyser frames [0, 2048) contains only the constant 1. 
PASS   At time 0.1328125 Analyser frames [0, 1920) contains only the constant 1. 
PASS   At time 0.1328125 Analyser frames [1920, 2048) contains only the constant 0. 
PASS   At time 0.140625 Analyser frames [0, 1792) contains only the constant 1. 
FAIL X At time 0.140625 Analyser frames [1792, 2048): Expected 0 for all values but found 128 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 124 more errors. assert_true: expected true got false
PASS   At time 0.1484375 Analyser frames [0, 1664) contains only the constant 1. 
FAIL X At time 0.1484375 Analyser frames [1664, 2048): Expected 0 for all values but found 256 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 252 more errors. assert_true: expected true got false
PASS   At time 0.15625 Analyser frames [0, 1536) contains only the constant 1. 
FAIL X At time 0.15625 Analyser frames [1536, 2048): Expected 0 for all values but found 384 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 380 more errors. assert_true: expected true got false
PASS   At time 0.1640625 Analyser frames [0, 1408) contains only the constant 1. 
FAIL X At time 0.1640625 Analyser frames [1408, 2048): Expected 0 for all values but found 512 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 508 more errors. assert_true: expected true got false
PASS   At time 0.171875 Analyser frames [0, 1280) contains only the constant 1. 
FAIL X At time 0.171875 Analyser frames [1280, 2048): Expected 0 for all values but found 640 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 636 more errors. assert_true: expected true got false
PASS   At time 0.1796875 Analyser frames [0, 1152) contains only the constant 1. 
FAIL X At time 0.1796875 Analyser frames [1152, 2048): Expected 0 for all values but found 768 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 764 more errors. assert_true: expected true got false
PASS   At time 0.1875 Analyser frames [0, 1024) contains only the constant 1. 
FAIL X At time 0.1875 Analyser frames [1024, 2048): Expected 0 for all values but found 896 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 892 more errors. assert_true: expected true got false
PASS   At time 0.1953125 Analyser frames [0, 896) contains only the constant 1. 
FAIL X At time 0.1953125 Analyser frames [896, 2048): Expected 0 for all values but found 1024 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1020 more errors. assert_true: expected true got false
PASS   At time 0.203125 Analyser frames [0, 768) contains only the constant 1. 
FAIL X At time 0.203125 Analyser frames [768, 2048): Expected 0 for all values but found 1152 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1148 more errors. assert_true: expected true got false
PASS   At time 0.2109375 Analyser frames [0, 640) contains only the constant 1. 
FAIL X At time 0.2109375 Analyser frames [640, 2048): Expected 0 for all values but found 1280 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1276 more errors. assert_true: expected true got false
PASS   At time 0.21875 Analyser frames [0, 512) contains only the constant 1. 
FAIL X At time 0.21875 Analyser frames [512, 2048): Expected 0 for all values but found 1408 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1404 more errors. assert_true: expected true got false
PASS   At time 0.2265625 Analyser frames [0, 384) contains only the constant 1. 
FAIL X At time 0.2265625 Analyser frames [384, 2048): Expected 0 for all values but found 1536 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1532 more errors. assert_true: expected true got false
PASS   At time 0.234375 Analyser frames [0, 256) contains only the constant 1. 
FAIL X At time 0.234375 Analyser frames [256, 2048): Expected 0 for all values but found 1664 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1660 more errors. assert_true: expected true got false
PASS   At time 0.2421875 Analyser frames [0, 128) contains only the constant 1. 
FAIL X At time 0.2421875 Analyser frames [128, 2048): Expected 0 for all values but found 1792 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1788 more errors. assert_true: expected true got false
FAIL X At time 0.25 Analyser frames [0, 2048): Expected 0 for all values but found 1920 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 1916 more errors. assert_true: expected true got false
FAIL < [timing] 15 out of 32 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 3 out of 3 tasks were failed. assert_true: expected true got false
